started to scaffold the climber subsystem

+++ src/main/java/frc/team1126/subsystems/sensors/LimelightHelpers.java
@@ -641,7 +641,7 @@ public class LimelightHelpers {
     /////
     /////
 
-    public static void setPipelineIndex(String limelightName, int pipelineIndex) {
+    public static void setPipelineIndex(String limelightName, double pipelineIndex) {
         setLimelightNTDouble(limelightName, "pipeline", pipelineIndex);
     }
